requirements.txt
python-dotenv
requests
fastapi
uvicorn
pytest
